// src/library/interfaces/ChampionMasteryDto.ts
export interface ChampionMasteryDto {
  championPointsUntilNextLevel: number;
  chestGranted: boolean;
  championId: number;
  lastPlayTime: number;
  championLevel: number;
  summonerId: string;
  championPoints: number;
  championPointsSinceLastLevel: number;
  tokensEarned: number;
  summonerLevel: number;
}